feat: add --debug mode
This commit adds a `--debug` mode to the `auto-commit` command.
When this mode is enabled, the command will log the commands it runs to the console.
This can be useful for debugging purposes or for understanding how the command works.

To use the `--debug` mode, simply pass the `--debug` flag to the `auto-commit` command.
For example:

```
auto-commit --debug
```