feat(interactive-messages): handle skipping input
This commit introduces a breaking change to the interactive commit message generator.

Previously, the generator would continue prompting the user for input until they entered a valid commit message. This could be frustrating for users who wanted to skip the input process and simply commit their changes with the default message.

This change introduces a new behavior where the generator will break out of the input loop if the user enters any non-empty input. This allows users to skip the input process by simply entering any character or string.

**Breaking Changes:**

- The `do { ... } while (false)` loop has been changed to `do { ... } while (true)`.
- The `if (ok)` conditional has been changed to `if (ok === "n" || ok === "no" || ok === "N" || ok === "NO")`.
- A new `else` clause has been added to the `if` conditional.